pyproject.toml
[project]
name = "algoforge"
version = "0.1.0"
description = "Hierarchical multi-agent system for automated algorithm discovery"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.scripts]
algoforge = "algoforge.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/algoforge"]

[tool.hatch.envs.default]
features = ["dev"]
